requirements.txt
Flask==3.0.3
flask-cors==4.0.1

requests==2.32.3
python-dotenv==1.0.1

google-generativeai==0.7.2